have done some error handling in user services

+++ server/src/services/user-service.js
@@ -15,6 +15,9 @@ async function createUser(data) {
   try {
     
     const gym = await gymRepository.findGym(data.gymId);
+    if(!gym) {
+      throw new AppError('The gym with the given gymId doesnt exist',StatusCodes.BAD_REQUEST);
+    }
     // console.log('gym : ', gym);
     const user = await userRepository.create({
             email : data.email,
@@ -57,6 +60,7 @@ async function createUser(data) {
     return {jwt, user};
   } catch (error) {
     // console.log(error);
+    if(error instanceof AppError) throw error;
     throw new AppError(
       'Cannot create a new User object',
       StatusCodes.INTERNAL_SERVER_ERROR
@@ -67,9 +71,14 @@ async function createUser(data) {
 async function getUser(id) {
   try {
     const user = await userRepository.get(id);
+    if(!user) {
+      throw new AppError('No User exists for the given UserId',StatusCodes.BAD_REQUEST);
+    }
     return user;
+   
   } catch (error) {
     // console.log(error);
+    if( error instanceof AppError) throw error;
     throw new AppError(
       'Cannot fetch data of the user',
       StatusCodes.INTERNAL_SERVER_ERROR
@@ -133,6 +142,7 @@ async function isAuthenticated(token) {
     if (error.name === 'TokenExpiredError') {
       throw new AppError('JWT token expired', StatusCodes.BAD_REQUEST);
     }
+    if(error instanceof AppError) throw error;
     throw new AppError(
       'Something went wrong',
       StatusCodes.INTERNAL_SERVER_ERROR
@@ -159,6 +169,7 @@ async function isAuthenticatedReset(id, token) {
     if (error.name === 'TokenExpiredError') {
       throw new AppError('JWT token expired', StatusCodes.BAD_REQUEST);
     }
+    if(error instanceof AppError) throw error;
     throw new AppError(
       'Something went wrong',
       StatusCodes.INTERNAL_SERVER_ERROR
@@ -171,9 +182,13 @@ async function getUserByUserId(id) {
     // console.log(id);
     const user = await userRepository.getUserByUserId(id);
     // console.log(user);
+    if(!user) {
+      throw new AppError('no user exist for this userId',StatusCodes.BAD_REQUEST);
+    }
     return user;
   } catch (error) {
     // console.log(error);
+    if(error instanceof AppError) throw error;
     throw new AppError(
       'Cannot fetch data of the user',
       StatusCodes.INTERNAL_SERVER_ERROR
@@ -184,10 +199,14 @@ async function getUserByUserId(id) {
 async function getUserInfo(data) {
   try {
     const user = await userRepository.getUserByNameAndGym(data);
+    if(!user) {
+      throw new AppError('no user exists for these values', StatusCodes.BAD_REQUEST);
+    }
     // console.log(user);
     return user;
   } catch (error) {
     // console.log(error);
+    if(error instanceof AppError) throw error;
     throw new AppError(
       'Cannot fetch data of the user',
       StatusCodes.INTERNAL_SERVER_ERROR
@@ -198,11 +217,21 @@ async function getUserInfo(data) {
 async function deleteUser(id) {
   try {
     const user = await userRepository.destroy(id);
+    if(!user) {
+      throw new AppError('no user exist for the given userId',StatusCodes.BAD_REQUEST);
+    }
     const gym = await gymRepository.deleteMembersFromGym(user.gymId, user._id);
+    if(gym.nModified === 0) {
+      throw new AppError('no such user exist in the gym members array',StatusCodes.BAD_GATEWAY);
+    }
     const userProfile = await userProfileRespository.deleteUserProfileByUserId(user._id);
+    if(!userProfile) {
+      throw new AppError('no userProfile exist for this user');
+    }
     const attendance = await  attendanceRespository.deleteAllAttendanceOfTheUserId(userProfile.attendance);
     return user;
   } catch (error) {
+    if( error instanceof AppError) throw error;
     throw new AppError(
       'Cannot delete user from the database',
       StatusCodes.INTERNAL_SERVER_ERROR
@@ -213,6 +242,9 @@ async function deleteUser(id) {
 async function addRoleToUser(id, data) {
   try {
     const response = await userRepository.update(id, data);
+    if(!response) {
+      throw new AppError('no user found for this userId', StatusCodes.BAD_REQUEST);
+    }
     return response;
   } catch (error) {
     throw new AppError(
